backend/requirements.txt
# -------------------------
# FastAPI & Core
# -------------------------
fastapi>=0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
python-dotenv==1.0.1

# -------------------------
# Security & Auth
# -------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1

# -------------------------
# Pydantic / Validation
# -------------------------
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic-extra-types==2.10.3
email-validator==2.2.0

# -------------------------
# Database
# -------------------------
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# -------------------------
# Background Tasks & Cache
# -------------------------
redis==5.2.1
celery==5.4.0
apscheduler==3.10.4

# -------------------------
# HTTP / Async
# -------------------------
httpx==0.28.1
aiohttp==3.11.10
websockets==14.1

# -------------------------
# Notifications
# -------------------------
twilio==9.3.7
sendgrid==6.11.0
firebase-admin==6.6.0

# -------------------------
# AI / ML
# -------------------------
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
imbalanced-learn==0.12.4
joblib==1.4.2
xgboost==2.1.3
lightgbm==4.5.0
tensorflow==2.18.0

# -------------------------
# Data / Time
# -------------------------
python-dateutil==2.9.0
pytz==2024.2

# -------------------------
# Logging & Monitoring
# -------------------------
loguru==0.7.3
prometheus-client==0.21.0

# -------------------------
# Testing
# -------------------------
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# -------------------------
# Visualization / Reports
# -------------------------
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1

# -------------------------
# Utilities
# -------------------------
python-slugify==8.0.4
shortuuid==1.0.13
